refactor: integrate chain_api
