+++ src/rpc/auth_layer.rs
@@ -3,7 +3,7 @@
 
 use crate::auth::{verify_token, JWT_IDENTIFIER};
 use crate::key_management::KeyStore;
-use crate::rpc::{auth_api, beacon_api, CANCEL_METHOD_NAME};
+use crate::rpc::{auth_api, beacon_api, chain_api, CANCEL_METHOD_NAME};
 use crate::rpc_api::*;
 use futures::future::BoxFuture;
 use futures::FutureExt;
