+++ src/rpc/mod.rs
@@ -5,7 +5,7 @@
 pub mod auth_api;
 mod auth_layer;
 pub mod beacon_api;
-mod chain_api;
+pub mod chain_api;
 mod channel;
 mod common_api;
 mod eth_api;
@@ -37,8 +37,8 @@ use crate::rpc::{
     state_api::*,
 };
 use crate::rpc_api::{
-    chain_api::*, common_api::*, eth_api::*, gas_api::*, mpool_api::*, net_api::*,
-    node_api::NODE_STATUS, state_api::*, sync_api::*, wallet_api::*,
+    common_api::*, eth_api::*, gas_api::*, mpool_api::*, net_api::*, node_api::NODE_STATUS,
+    state_api::*, sync_api::*, wallet_api::*,
 };
 
 use fvm_ipld_blockstore::Blockstore;
