+++ src/rpc_api/mod.rs
@@ -3,140 +3,6 @@
 
 /// JSON-RPC API definitions
 
-/// Chain API
-pub mod chain_api {
-    use std::{path::PathBuf, sync::Arc};
-
-    #[cfg(test)]
-    use crate::blocks::RawBlockHeader;
-    use crate::blocks::Tipset;
-    use crate::lotus_json::lotus_json_with_self;
-    #[cfg(test)]
-    use crate::lotus_json::{assert_all_snapshots, assert_unchanged_via_json};
-    use crate::lotus_json::{HasLotusJson, LotusJson};
-    use crate::rpc::types::ApiTipsetKey;
-    use crate::shim::clock::ChainEpoch;
-    use schemars::JsonSchema;
-    use serde::{Deserialize, Serialize};
-
-    pub const CHAIN_GET_MESSAGE: &str = "Filecoin.ChainGetMessage";
-
-    pub const CHAIN_EXPORT: &str = "Filecoin.ChainExport";
-
-    #[derive(Debug, Clone, Serialize, Deserialize)]
-    pub struct ChainExportParams {
-        pub epoch: ChainEpoch,
-        pub recent_roots: i64,
-        pub output_path: PathBuf,
-        #[serde(with = "crate::lotus_json")]
-        pub tipset_keys: ApiTipsetKey,
-        pub skip_checksum: bool,
-        pub dry_run: bool,
-    }
-
-    lotus_json_with_self!(ChainExportParams);
-
-    pub type ChainExportResult = Option<String>;
-
-    pub const CHAIN_READ_OBJ: &str = "Filecoin.ChainReadObj";
-    pub const CHAIN_HAS_OBJ: &str = "Filecoin.ChainHasObj";
-    pub const CHAIN_GET_BLOCK_MESSAGES: &str = "Filecoin.ChainGetBlockMessages";
-    pub const CHAIN_GET_TIPSET_BY_HEIGHT: &str = "Filecoin.ChainGetTipSetByHeight";
-    pub const CHAIN_GET_TIPSET_AFTER_HEIGHT: &str = "Filecoin.ChainGetTipSetAfterHeight";
-    pub const CHAIN_GET_GENESIS: &str = "Filecoin.ChainGetGenesis";
-    pub const CHAIN_HEAD: &str = "Filecoin.ChainHead";
-    pub const CHAIN_GET_BLOCK: &str = "Filecoin.ChainGetBlock";
-    pub const CHAIN_GET_TIPSET: &str = "Filecoin.ChainGetTipSet";
-    pub const CHAIN_GET_PATH: &str = "Filecoin.ChainGetPath";
-    pub const CHAIN_SET_HEAD: &str = "Filecoin.ChainSetHead";
-    pub const CHAIN_GET_MIN_BASE_FEE: &str = "Filecoin.ChainGetMinBaseFee";
-    pub const CHAIN_GET_MESSAGES_IN_TIPSET: &str = "Filecoin.ChainGetMessagesInTipset";
-    pub const CHAIN_GET_PARENT_MESSAGES: &str = "Filecoin.ChainGetParentMessages";
-    pub const CHAIN_NOTIFY: &str = "Filecoin.ChainNotify";
-    pub const CHAIN_GET_PARENT_RECEIPTS: &str = "Filecoin.ChainGetParentReceipts";
-
-    #[derive(PartialEq, Debug, Serialize, Deserialize, Clone, JsonSchema)]
-    #[serde(rename_all = "snake_case")]
-    pub enum PathChange<T = Arc<Tipset>> {
-        Revert(T),
-        Apply(T),
-    }
-    impl HasLotusJson for PathChange {
-        type LotusJson = PathChange<LotusJson<Tipset>>;
-
-        #[cfg(test)]
-        fn snapshots() -> Vec<(serde_json::Value, Self)> {
-            use serde_json::json;
-            vec![(
-                json!({
-                    "revert": {
-                        "Blocks": [
-                            {
-                                "BeaconEntries": null,
-                                "ForkSignaling": 0,
-                                "Height": 0,
-                                "Messages": { "/": "baeaaaaa" },
-                                "Miner": "f00",
-                                "ParentBaseFee": "0",
-                                "ParentMessageReceipts": { "/": "baeaaaaa" },
-                                "ParentStateRoot": { "/":"baeaaaaa" },
-                                "ParentWeight": "0",
-                                "Parents": [{"/":"bafyreiaqpwbbyjo4a42saasj36kkrpv4tsherf2e7bvezkert2a7dhonoi"}],
-                                "Timestamp": 0,
-                                "WinPoStProof": null
-                            }
-                        ],
-                        "Cids": [
-                            { "/": "bafy2bzaceag62hjj3o43lf6oyeox3fvg5aqkgl5zagbwpjje3ajwg6yw4iixk" }
-                        ],
-                        "Height": 0
-                    }
-                }),
-                Self::Revert(Arc::new(Tipset::from(RawBlockHeader::default()))),
-            )]
-        }
-
-        fn into_lotus_json(self) -> Self::LotusJson {
-            match self {
-                PathChange::Revert(it) => PathChange::Revert(LotusJson(Tipset::clone(&it))),
-                PathChange::Apply(it) => PathChange::Apply(LotusJson(Tipset::clone(&it))),
-            }
-        }
-
-        fn from_lotus_json(lotus_json: Self::LotusJson) -> Self {
-            match lotus_json {
-                PathChange::Revert(it) => PathChange::Revert(it.into_inner().into()),
-                PathChange::Apply(it) => PathChange::Apply(it.into_inner().into()),
-            }
-        }
-    }
-
-    #[cfg(test)]
-    impl<T> quickcheck::Arbitrary for PathChange<T>
-    where
-        T: quickcheck::Arbitrary,
-    {
-        fn arbitrary(g: &mut quickcheck::Gen) -> Self {
-            let inner = T::arbitrary(g);
-            g.choose(&[PathChange::Apply(inner.clone()), PathChange::Revert(inner)])
-                .unwrap()
-                .clone()
-        }
-    }
-
-    #[test]
-    fn snapshots() {
-        assert_all_snapshots::<PathChange>()
-    }
-
-    #[cfg(test)]
-    quickcheck::quickcheck! {
-        fn quickcheck(val: PathChange) -> () {
-            assert_unchanged_via_json(val)
-        }
-    }
-}
-
 /// Message Pool API
 pub mod mpool_api {
     pub const MPOOL_GET_NONCE: &str = "Filecoin.MpoolGetNonce";
